requirements.txt
flask
flask-socketio
flask-cors
pymongo[srv]
bcrypt
python-dotenv
openai
python-dateutil
gunicorn
eventlet
